Update data_logger2.py
the task.read in the for loop in line 47 returns a list of readings from all the channels in the device as each measurement; so here we are using enumeration to split this list of lists up into a list for each of the 8 channels separately.
For some reason it seems to go through the first device twice. It prints 'data' for Dev2, then crashes and indicates an index error.