Update data_logger2.py
the goal now is to be able to set the number of samples and the sampling rate to attain the accurate voltage readings.
For some reason, increasing the sampling rate and the number of samples in the time function drastically reduces the performance of the code.
A resource warning also appears saying that a task was not explicitly closed before being destructed.